refactor: move CQRS commands and queries to Features folder and namespace
https://heidari.atlassian.net/browse/CMS-10

- Moved all CQRS command classes to AlturaCMS.Application.Features.Commands namespace
- Moved all CQRS query classes to AlturaCMS.Application.Features.Queries namespace

This change reorganizes the CQRS structure by placing commands and queries into a more appropriate Features namespace, improving project structure and maintainability.
