+++ AlturaCMS.Application/Features/ContentTypes/Queries/GetContentTypeByIdQuery.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AlturaCMS.Application.ContentTypes.Queries;
+namespace AlturaCMS.Application.Features.ContentTypes.Queries;
 internal class GetContentTypeByIdQuery
 {
 }
