+++ AlturaCMS.Application/Features/Forms/Commands/CreateFormCommand.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AlturaCMS.Application.Forms.Commands;
+namespace AlturaCMS.Application.Features.Forms.Commands;
 internal class CreateFormCommand
 {
 }
